components renamed and moved to components folder
